==> work/00_init.R <==
library(svMisc)
library(velox)
library(randomForest)
library(census.tools)
library(raster)
library(dplyr)
library(ggplot2)
library(scales)
library(gridExtra)

data_path = "~/BurkeLab Dropbox/Data/satellite_review_paper"
git_path = "~/Documents/GitHub/satellite-review"
dhs_path = "~/BurkeLab Dropbox/dhs"
census_data_path = "~/Documents/data/census-groundtruth"
burke_census_path = "~/BurkeLab Dropbox/Anne Driscoll/AfricaPovertyPaper"
dropbox_path = "~/BurkeLab Dropbox/"
